// apps/trade-pro/src/pages/defineDivision/forms/division/types.ts
export type TDefineDivisionDataOnAdd = {
  CreatedOn: string;
  SortNo: number;
  BranchId: number;
  CompanyId: number;
  CreatedUserId: number;
  DivisionId: number;
  OrganizationId: number;
  StateProvinceId: number;
  DivisionCode: string;
  DivisionName: string;
};
export type TDefineTehsilDataOnAdd = {
  OrganizationId: number;
  CreatedOn: string;
  SortNo: number;
  BranchId: number;
  CompanyId: number;
  CreatedUserId: number;
  DistrictId: number;
  TehsilCode: string;
  TehsilName: string;
  TehsilId: number;
};